+++ crates/agglayer-aggregator-notifier/src/lib.rs
@@ -4,11 +4,10 @@ use std::sync::Arc;
 
 use agglayer_certificate_orchestrator::{Certifier, CertifierOutput, EpochPacker, Error};
 use agglayer_config::certificate_orchestrator::prover::ProverConfig;
+use agglayer_types::Certificate;
 use error::NotifierError;
 use futures::future::BoxFuture;
-use pessimistic_proof::{
-    certificate::Certificate, local_state::LocalNetworkStateData, LocalNetworkState,
-};
+use pessimistic_proof::{local_state::LocalNetworkStateData, LocalNetworkState};
 use proof::Proof;
 use reth_primitives::Address;
 use serde::Serialize;
