+++ crates/agglayer-certificate-orchestrator/Cargo.toml
@@ -12,6 +12,7 @@ tokio-util.workspace = true
 tracing.workspace = true
 
 agglayer-clock = { path = "../agglayer-clock" }
+agglayer-types = { path = "../agglayer-types" }
 pessimistic-proof = { path = "../pessimistic-proof" }
 tokio-stream = { version = "0.1.15", features = ["sync"] }
 futures-util = "0.3.30"
