+++ crates/agglayer-node/Cargo.toml
@@ -29,6 +29,7 @@ tracing.workspace = true
 agglayer-config = { path = "../agglayer-config" }
 agglayer-clock = { path = "../agglayer-clock" }
 agglayer-telemetry = { path = "../agglayer-telemetry" }
+agglayer-types = { path = "../agglayer-types" }
 agglayer-signer = { path = "../agglayer-signer" }
 agglayer-certificate-orchestrator = { path = "../agglayer-certificate-orchestrator" }
 pessimistic-proof = { path = "../pessimistic-proof" }
